+++ tests/I18n/I18nTest.php
@@ -23,4 +23,12 @@ class I18nTest extends TestCase
         $name = Yii::$app->getI18n()->getTableName('test', 'de');
         self::assertEquals('{{%test_de}}', $name);
     }
+
+    public function testLabels(): void
+    {
+        $i18n = Yii::$app->getI18n();
+
+        self::assertSame('Deutsch', $i18n->getLabel('de'));
+        self::assertSame('nl', $i18n->getLabel('nl'));
+    }
 }